Splitting dto file up
