+++ source/dotnet/Services/Application/Application.csproj
@@ -33,8 +33,8 @@ limitations under the License.
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Include="..\..\Contracts\BatchGridAreaStepsDto.cs">
-      <Link>Batches\BatchGridAreaStepsDto.cs</Link>
+    <Compile Include="..\..\Contracts\BatchGridAreaDto.cs">
+      <Link>Batches\BatchGridAreaDto.cs</Link>
     </Compile>
     <Compile Include="..\..\Contracts\BatchRequestDto.cs">
       <Link>Batches\BatchRequestDto.cs</Link>
@@ -44,6 +44,18 @@ limitations under the License.
     </Compile>
     <Compile Include="..\..\Contracts\GridAreaDto.cs">
       <Link>Batches\GridAreaDto.cs</Link>
+    </Compile>
+    <Compile Include="..\..\Contracts\MeteringPointType.cs">
+      <Link>Batches\MeteringPointType.cs</Link>
+    </Compile>
+    <Compile Include="..\..\Contracts\ProcessDto.cs">
+      <Link>Batches\ProcessDto.cs</Link>
+    </Compile>
+    <Compile Include="..\..\Contracts\ProcessMetaDataDto.cs">
+      <Link>Batches\ProcessMetaDataDto.cs</Link>
+    </Compile>
+    <Compile Include="..\..\Contracts\TimeSeriesPointDto.cs">
+      <Link>Batches\TimeSeriesPointDto.cs</Link>
     </Compile>
     <Compile Include="..\..\Contracts\WholesaleProcessType.cs">
       <Link>Processes\WholesaleProcessType.cs</Link>
